refactor(redux): split actions, reducers and sagas into few modules
