+++ src/components/PrivateRoute.jsx
@@ -3,7 +3,7 @@ import { Route, Redirect } from "react-router-dom";
 import { useSelector } from "react-redux";
 
 const PrivateRoute = ({ path, component: RouteComponent }) => {
-  const { loginStatus } = useSelector((state) => state);
+  const { loginStatus } = useSelector((state) => state.auth);
 
   return (
     <>
