+++ src/index.js
@@ -7,7 +7,7 @@ import { MuiThemeProvider } from "@material-ui/core/styles";
 import { BrowserRouter } from "react-router-dom";
 import reportWebVitals from "./reportWebVitals";
 import { Provider } from "react-redux";
-import store from "./redux/store";
+import store from "./store";
 
 ReactDOM.render(
   <React.StrictMode>
