+++ src/modules/route/index.js
@@ -1,0 +1,3 @@
+export { default } from "./reducer";
+export * from "./actions";
+export * from "./saga";
