--- src/redux/constants.js
@@ -1,35 +1,0 @@
-export const LOGIN_REQUEST = "LOGIN_REQUEST";
-export const LOGIN_REQUEST_SUCCESSFUL = "LOGIN_REQUEST_SUCCESSFUL";
-export const LOGIN_REQUEST_FAILED = "LOGIN_REQUEST_FAILED";
-
-export const REGISTRATION_REQUEST = "REGISTRATION_REQUEST";
-export const REGISTRATION_REQUEST_SUCCESSFUL =
-  "REGISTRATION_REQUEST_SUCCESSFUL";
-export const REGISTRATION_REQUEST_FAILED = "REGISTRATION_REQUEST_FAILED";
-
-export const POST_CARD_REQUEST = "POST_CARD_REQUEST";
-export const POST_CARD_REQUEST_SUCCESSFUL = "POST_CARD_REQUEST_SUCCESSFUL";
-export const POST_CARD_REQUEST_FAILED = "POST_CARD_REQUEST_FAILED";
-
-export const GET_CARD_REQUEST = "GET_CARD_REQUEST";
-export const GET_CARD_REQUEST_SUCCESSFUL = "GET_CARD_REQUEST_SUCCESSFUL";
-export const GET_CARD_REQUEST_FAILED = "GET_CARD_REQUEST_FAILED";
-
-export const LOGOUT_BUTTON_PRESSED = "LOGOUT_BUTTON_PRESSED";
-export const LOGOUT = "LOGOUT";
-
-export const LOADING_START = "LOADING_START";
-export const LOADING_DONE = "LOADING_DONE";
-
-export const PROFILE_OPEN = "PROFILE_OPEN";
-export const PROFILE_CLOSE = "PROFILE__CLOSE";
-
-export const ADDRESSES_LIST_REQUEST = "ADDRESSES_LIST_REQUEST";
-export const ADDRESSES_LIST_REQUEST_SUCCESSFUL =
-  "ADDRESSES_LIST_REQUEST_SUCCESSFUL";
-export const ADDRESSES_LIST_REQUEST_FAILED = "ADDRESSES_LIST_REQUEST_FAILED";
-
-export const ROUTE_REQUEST = "ROUTE_REQUEST";
-export const ROUTE_REQUEST_SUCCESSFUL = "ROUTE_REQUEST_SUCCESSFUL";
-export const ROUTE_REQUEST_FAILED = "ROUTE_REQUEST_FAILED";
-export const ROUTE_RESET = "ROUTE_RESET";
